:seedling: Create the vsix package from `dist/`
Explicitly copy all files needed by the vscode extension
to a `dist/` directory and build the `.vsix` file from
there.

There are a few advantages to using this method:

  - Only what we explicitly copy to the `dist/` directory
    will be included in the `.vsix` file

  - Every developer and CI script will be able to use the
    same packaging scripts to maintain consistency

  - Changes to what is included is centralized in a
    few scripts

Note: In hopes of easier build support on windows, the scripts
are all javascript/nodejs.  Getting bash scripts to be stable
on windows can be a challenge, and maintaining platform specific
scripts is best avoided if possible.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>