:seedling: Include dummy `GENAI_KEY` in rpc server env
The kai rpc server needed the `GENAI_KEY` envvar to be
set so it can start up.  A valid key is _not_ needed to
be able to run analysis.

Update the kai rpc server env function to insert the
dummy value but let it be overridden by the value available
in vscode's the actual envvars.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>