:sparkles: Enhance server process and status handling
Resolves: #133
Resolves: #135

Use additional child_process eventing to keep track of the
server process itself.  This lets the extension time the
initialize call better, and be able to "cleanly" shutdown
the `kai-rpc-server``.  If the `kai-rpc-server` is killed or
terminated externally, it will be noticed and the extension
will understand the server is stopped.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>