:seedling: Update collect-assets scripts
Resolves: #234
Resolves: #334

Update the collect-assets script such that assets
can be sourced from GitHub release assess or from
a GitHub workflow's artifacts.  Download of assets
and artifacts is separated from unpacking of
the assets.

When complete, the full set of downloaded assets
that the editor-extensions can use will be located
under `{project_root}/downloaded_assets`.

Downloaded files will have the ETag stored in a
sidecar file.  This allows future runs to skip the
download step if the url's data has not been
changed.

Note: File globs are used during unpacking to allow
extracting only a portion of an archive that is
relevant to the desired file set.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>