:seedling: drop repo asset fernflower
The fernflower java decompiler is not referenced
or sent to the Kai analysis server at this time.
Drop the jar from the repo, but leave the path
references in place.  If it is needed in the future,
a download step can be added to `collect-assets.js`,
and a copy step added to `copy-dist.js`.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>