:seedling: drop repo asset `jdtls-bundles/java-analyzer-bundle`
The `jdtls-bundles/java-analyzer-bundle` jars may be
downloaded and extracted from one of the kai release
assets. The extra work to extract this file is done in #453.

Drop the files from the repo and adjust the asset paths
to reference at dev time and copy at dist time from the
`downloaded_assets/` folder.

Follows: #453
Partial: #234

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>