Adding helper methods to the proxy, to convert from vscode objects
When the proxy makes the vscode commands to proxy the requests return
vscode objects, not the LSP protocol versions. We have to convert them.
There are helpers, here:
https://github.com/microsoft/vscode-languageserver-node/blob/cd32d3e505e8ec8df660bc8e328d5f650673c2d8/client/src/common/codeConverter.ts#L936-L943
but I can not figure out how to get that converter.

If you know how, feel free to either fix in this PR or as a follow up.

Signed-off-by: Shawn Hurley <shawn@hurley.page>